Experimental code to add parsing support to envy.
This adds envy_parse, with two functions related to parsing host/ip
addresses.

to_ip/2 takes a host/ip address strings, parses, validates it, and
returns it as an IPv4 or IPv6 tuple based on a preferences array.

host_to_ip/2 extracts a key from the config, and parses it according
to the preferences specified in the ip_mode key.